+++ frontends/torch-frontend/CMakeLists.txt
@@ -5,6 +5,9 @@ if(POLICY CMP0116)
 endif()
 
 project(torch-frontend)
+
+option(TORCH_FRONTEND_ENABLE_JIT_IR_IMPORTER "Enable JIT IR Importer" ON)
+
 set(TORCH_FRONTEND_SRC_ROOT ${CMAKE_CURRENT_SOURCE_DIR})
 set(TORCH_FRONTEND_BIN_ROOT ${CMAKE_CURRENT_BINARY_DIR})
