+++ frontends/torch-frontend/TorchMLIR.cmake
@@ -22,8 +22,9 @@ function(build_torch_mlir)
     -DCMAKE_VISIBILITY_INLINES_HIDDEN=ON
     -DCMAKE_C_VISIBILITY_PRESET=hidden
     -DCMAKE_CXX_VISIBILITY_PRESET=hidden
+    -DTORCH_MLIR_ENABLE_REFBACKEND=OFF
     -DTORCH_MLIR_ENABLE_LTC=OFF
-    -DTORCH_MLIR_ENABLE_PYTORCH_EXTENSIONS=ON
+    -DTORCH_MLIR_ENABLE_PYTORCH_EXTENSIONS=${TORCH_FRONTEND_ENABLE_JIT_IR_IMPORTER}
     -DCMAKE_CXX_FLAGS=${CMAKE_CXX_FLAGS}
     -DPython3_EXECUTABLE=${Python3_EXECUTABLE}
     -DLLVM_EXTERNAL_LIT=${LLVM_EXTERNAL_LIT}
