+++ frontends/torch-frontend/torch-frontend/python/version.txt
@@ -1,1 +1,1 @@
-1.0.7
+1.1.0
